+++ Changelog.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added [QuestEval](models/scialom2021/Readme.md)
 - Added [MoverScore](models/zhao2019/Readme.md)
 
+### Changed
+- Changed the QAEval interface to match other text generation metrics.
+The backend was also changed to not rely on SacreROUGE.
+
 ## [v0.0.2](https://github.com/danieldeutsch/repro/repro/tag/0.0.2) - 2021-07-30
 ### Added
 - Added a `RecipeGenerationModel` class